fixes on web app, new design

+++ src/app/Console/Commands/CleanFolders.php
@@ -8,20 +8,20 @@ use Illuminate\Support\Facades\Storage;
 class CleanFolders extends Command
 {
 
-    protected $signature = 'factotum:clean-laravel-scaffolding';
+	protected $signature = 'factotum:clean-laravel-scaffolding';
 
 
-    protected $description = 'Clean the default laravel scaffolding files and folders';
+	protected $description = 'Clean the default laravel scaffolding files and folders';
 
 
-    public function __construct()
-    {
-        parent::__construct();
-    }
+	public function __construct()
+	{
+		parent::__construct();
+	}
 
 
-    public function handle()
-    {
+	public function handle()
+	{
 		$this->_cleanUpWebRoutes();
 		$this->_cleanUpApiRoutes();
 		$this->_cleanUpMigrations();
@@ -30,7 +30,7 @@ class CleanFolders extends Command
 		$this->_cleanUpViews();
 		$this->_cleanUpSass();
 		$this->_cleanUpJs();
-    }
+	}
 
 
 	private function _cleanUpFolder( $folder )
@@ -65,7 +65,7 @@ class CleanFolders extends Command
 	}
 
 
-    private function _cleanUpMigrations()
+	private function _cleanUpMigrations()
 	{
 		$this->_cleanUpFolder( database_path( 'migrations') );
 	}
@@ -94,14 +94,18 @@ class CleanFolders extends Command
 	private function _cleanUpSass()
 	{
 		$this->_cleanUpFolder( resource_path( 'sass') );
-		rmdir( resource_path( 'sass') );
+		if ( file_exists(resource_path( 'sass')) ) {
+			rmdir( resource_path( 'sass') );
+		}
 	}
 
 
 	private function _cleanUpJs()
 	{
 		$this->_cleanUpFolder( resource_path( 'js') );
-		rmdir( resource_path( 'js') );
+		if ( file_exists(resource_path( 'js')) ) {
+			rmdir( resource_path( 'js') );
+		}
 	}
 
 }
